+++ packages/react-native-gesture-handler/src/v3/NativeDetector/NativeDetector.tsx
@@ -4,7 +4,7 @@ import HostGestureDetector from './HostGestureDetector';
 import { tagMessage } from '../../utils';
 import { LogicChildren, LogicMethods, Gesture } from '../types';
 import { invokeDetectorEvent } from '../hooks/utils';
-import { DetectorContext } from '../useDetectorContext';
+import { DetectorContext } from './useDetectorContext';
 import { Reanimated } from '../../handlers/gestures/reanimatedWrapper';
 import { configureRelations } from './utils';
 import { isComposedGesture } from '../hooks/utils/relationUtils';
